refactor(#30): delete redundante /zerowaste directory
