+++ backend/zerowaste/src/main/java/com/zerowaste/ZerowasteApplication.java
@@ -1,0 +1,13 @@
+package com.zerowaste;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ZerowasteApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ZerowasteApplication.class, args);
+	}
+
+}
