+++ backend/zerowaste/src/main/java/com/zerowaste/config/README.md
