+++ backend/zerowaste/src/main/java/com/zerowaste/config/auth/SpringSecurityConfiguration.java
@@ -1,0 +1,50 @@
+package com.zerowaste.config.auth;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.HttpMethod;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+import com.zerowaste.services.users.AuthenticateUserService;
+
+@Configuration
+@EnableWebSecurity
+public class SpringSecurityConfiguration {
+    @Autowired
+    private AuthenticateUserService authenticationUserService;
+
+    @Bean
+    public SecurityFilterChain securityFilterChain(HttpSecurity httpSecurity) throws Exception {
+        return httpSecurity
+        .csrf(csrf -> csrf.disable())
+        .cors(cors -> cors.disable())
+        .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
+        .authorizeHttpRequests(authorize -> authorize
+            .requestMatchers(HttpMethod.POST, "/users", "/users/login").permitAll()
+            .requestMatchers(HttpMethod.OPTIONS, "/users", "/users/login").permitAll()
+            .requestMatchers(HttpMethod.GET, "/users/check-auth-token").hasRole("USER")
+            .anyRequest().authenticated()
+        )
+        .addFilterBefore(authenticationUserService, UsernamePasswordAuthenticationFilter.class)
+        .build();
+    }
+
+    @Bean
+    public AuthenticationManager authenticationManager(AuthenticationConfiguration authenticationConfiguration) throws Exception {
+        return authenticationConfiguration.getAuthenticationManager();
+    }
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+}
