+++ backend/zerowaste/src/main/java/com/zerowaste/models/user/UserRole.java
@@ -1,0 +1,16 @@
+package com.zerowaste.models.user;
+
+public enum UserRole {
+    USER("USER"),
+    ADMIN("ADMIN");
+
+    private final String role;
+
+    UserRole(String role) {
+        this.role = role;
+    }
+
+    public String getRole() {
+        return role;
+    }
+}
