+++ backend/zerowaste/src/main/java/com/zerowaste/repositories/UsersRepository.java
@@ -1,0 +1,9 @@
+package com.zerowaste.repositories;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.zerowaste.models.user.User;
+
+public interface UsersRepository extends JpaRepository<User, Long> {
+    public User findByEmail(String email);
+}
