+++ backend/zerowaste/src/main/java/com/zerowaste/services/users/exceptions/UserWithSameEmailAlreadyExistsException.java
@@ -1,0 +1,7 @@
+package com.zerowaste.services.users.exceptions;
+
+public class UserWithSameEmailAlreadyExistsException extends RuntimeException {
+    public UserWithSameEmailAlreadyExistsException(String message) {
+        super(message);
+    }
+}
