--- backend/zerowaste/src/main/java/com/zerowaste/zerowaste/ZerowasteApplication.java
@@ -1,13 +1,0 @@
-package com.zerowaste.zerowaste;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ZerowasteApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ZerowasteApplication.class, args);
-	}
-
-}
