--- backend/zerowaste/src/main/java/com/zerowaste/zerowaste/controllers/HealthCheckController.java
@@ -1,21 +1,0 @@
-package com.zerowaste.zerowaste.controllers;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController()
-public class HealthCheckController {
-    @GetMapping("/health")
-    public ResponseEntity<Map<String, String>> healthCheck() {
-        Map<String, String> responseBody = new HashMap<String, String>();
-
-        responseBody.put("message", "Hello World!");
-
-        return ResponseEntity.status(HttpStatus.OK).body(responseBody);
-    }
-}
