--- backend/zerowaste/src/main/java/com/zerowaste/zerowaste/repositories/UsersRepository.java
@@ -1,9 +1,0 @@
-package com.zerowaste.zerowaste.repositories;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.zerowaste.zerowaste.models.user.User;
-
-public interface UsersRepository extends JpaRepository<User, Long> {
-    public User findByEmail(String email);
-}
