--- backend/zerowaste/src/main/java/com/zerowaste/zerowaste/services/users/exceptions/UserWithSameEmailAlreadyExistsException.java
@@ -1,7 +1,0 @@
-package com.zerowaste.zerowaste.services.users.exceptions;
-
-public class UserWithSameEmailAlreadyExistsException extends RuntimeException {
-    public UserWithSameEmailAlreadyExistsException(String message) {
-        super(message);
-    }
-}
