--- backend/zerowaste/src/main/java/com/zerowaste/zerowaste/utils/validation/ValidEnum/ValidEnum.java
@@ -1,22 +1,0 @@
-package com.zerowaste.zerowaste.utils.validation.ValidEnum;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-import jakarta.validation.Constraint;
-import jakarta.validation.Payload;
-
-@Target({ ElementType.METHOD, ElementType.FIELD })
-@Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = EnumValidator.class)
-public @interface ValidEnum {
-    String message() default "must be any of enum {targetClassType}";
-
-    Class<?>[] groups() default {};
-
-    Class<? extends Payload>[] payload() default {};
-
-    Class<? extends Enum<?>> targetClassType();
-}
