--- backend/zerowaste/src/main/java/com/zerowaste/zerowaste/utils/validation/ValidationExceptionFilter.java
@@ -1,27 +1,0 @@
-package com.zerowaste.zerowaste.utils.validation;
-
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.FieldError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class ValidationExceptionFilter {
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<Map<String, List<String>>> handleValidationErrors(MethodArgumentNotValidException ex) {
-        Map<String, List<String>> errors = ex
-                .getBindingResult()
-                .getFieldErrors()
-                .stream()
-                .collect(Collectors.groupingBy(FieldError::getField,
-                        Collectors.mapping(FieldError::getDefaultMessage, Collectors.toList())));
-
-        return new ResponseEntity<>(errors, HttpStatus.BAD_REQUEST);
-    }
-}
